+++ DependencyInjection/OkaFileExtension.php
@@ -85,25 +85,40 @@ class OkaFileExtension extends Extension
 		$loader->load(sprintf('doctrine-behavior/%s.yml', $config['db_driver']));
 		$container->setParameter('oka_file.doctrine_behaviors.reflection.is_recursive', $config['behaviors']['reflection']['enable_recursive']);
 		
-		// PictureCoverizable behavior configuration
-		$pictureCoverizableListenerDefinition = $container->getDefinition('oka_file.doctrine_behaviors.picturecoverizable_listener');
-		
-		if ($config['behaviors']['picture_coverizable']['enabled'] === true) {
-			$pictureCoverizableListenerDefinition->replaceArgument(0, $config['behaviors']['picture_coverizable']['mappings']);
-			$pictureCoverizableListenerDefinition->replaceArgument(1, $config['object_default_class']['image']);
-		} else {
-			$pictureCoverizableListenerDefinition->clearTags();
+		foreach ($config['behaviors'] as $key => $behavior) {
+			if (true === in_array($key, ['reflection', 'enabled'], SORT_REGULAR)) {
+				continue;
+			}
+			
+			$behaviorListenerDefinition = $container->getDefinition(sprintf('oka_file.doctrine_behaviors.%s_listener', preg_replace('#_+#', '', $key)));
+			
+			if ($behavior['enabled'] === true) {
+				$behaviorListenerDefinition->replaceArgument(0, $behavior['mappings']);
+				$behaviorListenerDefinition->replaceArgument(1, $config['object_default_class']['image']);
+			} else {
+				$behaviorListenerDefinition->clearTags();
+			}
 		}
+
+// 		// PictureCoverizable behavior configuration
+// 		$pictureCoverizableListenerDefinition = $container->getDefinition('oka_file.doctrine_behaviors.picturecoverizable_listener');
 		
-		// Avatarizable behavior configuration
-		$avatarizableListenerDefinition = $container->getDefinition('oka_file.doctrine_behaviors.avatarizable_listener');
+// 		if ($config['behaviors']['picture_coverizable']['enabled'] === true) {
+// 			$pictureCoverizableListenerDefinition->replaceArgument(0, $config['behaviors']['picture_coverizable']['mappings']);
+// 			$pictureCoverizableListenerDefinition->replaceArgument(1, $config['object_default_class']['image']);
+// 		} else {
+// 			$pictureCoverizableListenerDefinition->clearTags();
+// 		}
 		
-		if ($config['behaviors']['avatarizable']['enabled'] === true) {
-			$avatarizableListenerDefinition->replaceArgument(0, $config['behaviors']['avatarizable']['mappings']);
-			$avatarizableListenerDefinition->replaceArgument(1, $config['object_default_class']['image']);
-		} else {
-			$avatarizableListenerDefinition->clearTags();
-		}
+// 		// Avatarizable behavior configuration
+// 		$avatarizableListenerDefinition = $container->getDefinition('oka_file.doctrine_behaviors.avatarizable_listener');
+		
+// 		if ($config['behaviors']['avatarizable']['enabled'] === true) {
+// 			$avatarizableListenerDefinition->replaceArgument(0, $config['behaviors']['avatarizable']['mappings']);
+// 			$avatarizableListenerDefinition->replaceArgument(1, $config['object_default_class']['image']);
+// 		} else {
+// 			$avatarizableListenerDefinition->clearTags();
+// 		}
 	}
 	
 	protected function loadFileStorageConfiguration(array $config, ContainerBuilder $container)