+++ DoctrineBehaviors/ODM/AbstractListener.php
@@ -15,7 +15,7 @@ abstract class AbstractListener extends BaseAbstractListener
 	 * @param array $mapping
 	 * @return array
 	 */
-	protected function handleEntityMapping($class, $mapping)
+	protected function handleDocumentMapping($class, $mapping)
 	{
 		if (isset($this->mappings[$class])) {
 			$mapping['targetDocument'] = isset($this->mappings[$class]['target_object']) ? $this->mappings[$class]['target_object'] :